fixed gcc compile warnings

+++ arangod/HashIndex/hasharray.c
@@ -124,7 +124,7 @@ static bool AllocateTable (TRI_hasharray_t* array, size_t numElements) {
   }
 
   // position array directly on a cache line boundary
-  offset = ((uint64_t) data) % CACHE_LINE_SIZE;
+  offset = ((intptr_t) data) % CACHE_LINE_SIZE;
 
   if (offset == 0) {
     // we're already on a cache line boundary
@@ -134,7 +134,7 @@ static bool AllocateTable (TRI_hasharray_t* array, size_t numElements) {
     // move to start of a cache line
     table = data + (CACHE_LINE_SIZE - offset);
   }
-  assert(((uint64_t) table) % CACHE_LINE_SIZE == 0);
+  assert(((intptr_t) table) % CACHE_LINE_SIZE == 0);
 
   array->_data = data;
   array->_table = table;
